% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance_matrix_knn_algorithm.R
\name{distance_matrix_banddepth}
\alias{distance_matrix_banddepth}
\title{Pairwise distance matrix based on the band depth distance}
\usage{
distance_matrix_banddepth(unit)
}
\arguments{
\item{unit}{A matrix representing a multivariate time series where each
column is a univariate time series.}
}
\value{
a matrix with pairwise distances
}
\description{
Pairwise distance matrix based on the band depth distance
}
\examples{
X=matrix( rnorm(2000), ncol=10  )
distance_matrix_banddepth(unit=X)
}
\seealso{
Band Depth Clustering for Nonstationary Time Series and
Wind Speed Behavior (2018) Tupper et al
}
