% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance_matrix_knn_algorithm.R
\name{distance_matrix_wasserstein}
\alias{distance_matrix_wasserstein}
\title{Distance matrix from based on the Wasserstein distance}
\usage{
distance_matrix_wasserstein(unit)
}
\arguments{
\item{unit}{A matrix representing a multivariate time series where each
column is a univariate time series.}
}
\value{
a matrix with pairwise distances
}
\description{
Pairwise distance matrix of a multivariate time series based on the
Wasserstein distance between the empirical distribution of the series
}
\examples{
X=matrix( rnorm(2000), ncol=10  )
distance_matrix_wasserstein(unit=X)
}
\seealso{
For more details, check the \code{transport} package documentation on CRAN
}
