% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerCausalityIndeces.R
\name{mBTS}
\alias{mBTS}
\title{modified Back-in-time Selection for vector AR parameters estimation}
\usage{
mBTS(xM, responseindex, pmax)
}
\arguments{
\item{xM}{the matrix of K time series (variables in columns)}

\item{responseindex}{the index of the response variable in {1,...,K}}

\item{ordersV}{vector of size 1xK of the maximum order for each of the K
variables.}

\item{indexV}{the vector of size 1 x K\emph{pmax of zeros and ones
e.g. if the component in position 2}pmax+3 is one, the
third variable, lag 3, X3(t-3), is selected.}
}
\value{
the matrix of all explanatory lagged variables in the
DR model. The sequence of the lagged variables in 'lagM'
}
\description{
modified Back-in-time Selection for vector AR parameters estimation
}
\seealso{
I. Vlachos and D. Kugiumtzis, “Backward-in-time selection of the
order of dynamic regression prediction model,” J. Forecast.,
vol. 32, pp. 685–701, 2013.
}
