% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerCausalityIndeces.R
\name{mBTS_Af_mat}
\alias{mBTS_Af_mat}
\title{mBTS Vector Autoregressive coefficients fourier transform}
\usage{
mBTS_Af_mat(xM, responseindex, pmax, freqs)
}
\arguments{
\item{xM}{the matrix of K time series (variables in columns)}

\item{responseindex}{the index of the response variable in \eqn{\{1,\ldots,K\}}}

\item{pmax}{maximum order(lag) of the VAR model to be considered}

\item{freqs}{frequencies at which the spectral density is estimated}
}
\value{
the matrix of all the Restricted Generalized Partial Directed
Coherence index across the series of a multivariate set.
}
\description{
DFT Vector Autoregressive coefficients matrix using the mBTS algorithm.
This matrix is the base to compute the generalized partial directed coherence
}
