% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerCausalityIndeces.R
\name{mBTSCGCI}
\alias{mBTSCGCI}
\title{computation of the conditional Granger causality index}
\usage{
mBTSCGCI(xM, responseindex, pmax)
}
\arguments{
\item{xM}{the matrix of K time series (variables in columns)}

\item{responseindex}{the index of the response variable in \eqn{\{1,\ldots,K\}}}

\item{pmax}{maximum order(lag) of the VAR model to be considered}
}
\value{
the matrix of all the conditional Granger causality index across
the series of a multivariate set.
}
\description{
computation of the conditional Granger causality index
}
\seealso{
Siggiridou, Elsa, and Dimitris Kugiumtzis. "Granger Causality
in Multivariate Time Series Using a Time-Ordered Restricted Vector
Autoregressive Model." IEEE Transactions on Signal Processing 64,
no. 7 (April 2016): 1759-73. \doi{10.1109/TSP.2015.2500893}
}
