% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDCmatrix.R
\name{sqnorms}
\alias{sqnorms}
\title{Quadratic multiplication of a matrix M with respect to a matrix A:
Conj(M)  A M, where Conj() is the complex conjugate function}
\usage{
sqnorms(M, A)
}
\arguments{
\item{M}{A Matrix of dimension P by P}

\item{A}{A squared Matrix of dimension P by P}
}
\value{
The squared root of the absolute values of the matrix result of the
operation Conj(M) A M
}
\description{
Quadratic multiplication of a matrix M with respect to a matrix A:
Conj(M)  A M, where Conj() is the complex conjugate function
}
\examples{
M=matrix( rnorm(100), ncol=10  )
A=matrix( rnorm(100), ncol=10  )
sqnorms(M, A)
}
